Extract logic from application controller to separate module
